refactor: extracted fetch logic in resuable hook
