+++ frontend/src/Pages/Dashboard/index.tsx
@@ -1,11 +1,9 @@
 import Box from '@mui/material/Box';
-import Container from '@mui/material/Container';
 import Grid from '@mui/material/Grid';
 import Paper from '@mui/material/Paper';
 import { QueueSize } from './Queue';
 import { ProgressSize } from './Progress';
 import Completed from './Completed';
-import Workflows from './Workflows';
 import Runners from './Runners';
 import GitHubStatus from './GitHubStatus';
 
@@ -16,14 +14,15 @@ function Dashboard() {
       sx={{
         backgroundColor: (theme) =>
           theme.palette.mode === 'light' ? theme.palette.grey[100] : theme.palette.grey[900],
-        flexGrow: 1,
-        height: '100vh',
-        overflow: 'auto',
+        maxWidth: '60rem',
+        margin: '0 auto',
+        padding: '0 1rem',
+        paddingBottom: '2rem',
       }}
     >
-      <Container maxWidth="lg" sx={{ mt: 4, mb: 4 }}>
-        <Grid container spacing={3}>
-          {/* Recent workflow queue */}
+      <Grid container spacing={2}>
+        {/* Recent workflow queue */}
+        <Grid item xs={12}>
           <Grid container spacing={2}>
             <Grid item xs={12} md={4} lg={3}>
               <Paper sx={{ p: 2, display: 'flex', flexDirection: 'column', height: 240 }}>
@@ -46,19 +45,14 @@ function Dashboard() {
               </Paper>
             </Grid>
           </Grid>
-          {/* Recent Workflows */}
-          <Grid item xs={12}>
-            <Paper sx={{ p: 2, display: 'flex', flexDirection: 'column' }}>
-              <Runners />
-            </Paper>
-          </Grid>
-          <Grid item xs={12}>
-            <Paper sx={{ p: 2, display: 'flex', flexDirection: 'column' }}>
-              <Workflows />
-            </Paper>
-          </Grid>
         </Grid>
-      </Container>
+        {/* Recent Workflows */}
+        <Grid item xs={12}>
+          <Paper sx={{ p: 2, display: 'flex', flexDirection: 'column' }}>
+            <Runners />
+          </Paper>
+        </Grid>
+      </Grid>
     </Box>
   );
 }
